Modify: scss 꺠짐현상오류 수정완료
