+++ src/Pages/Seokmoon/Main/Main.scss
@@ -127,7 +127,6 @@ li {
 }
 
 .main-image-box {
-  border: 1px solid red;
   width: 50px;
   height: 50px;
   border-radius: 70%;
@@ -143,7 +142,6 @@ li {
 }
 
 .main-feeds {
-  border: 1px solid red;
   width: 600px;
   height: 1200px;
   margin-left: 100px;
@@ -154,7 +152,6 @@ li {
 }
 
 .feed-nav {
-  border: 1px solid red;
   height: 70px;
   display: flex;
   background-color: $whiteColor;
@@ -257,6 +254,8 @@ input::placeholder {
   margin-top: 50px;
   width: 400px;
   height: 450px;
+  margin-bottom: 600px;
+  margin-left: 50px;
 }
 
 .right-box {
